+++ Sources/SwiftBeanCountWealthsimpleMapper/Extensions/Wealthsimple.Transaction+Helper.swift
