+++ Tests/SwiftBeanCountWealthsimpleMapperTests/WealthsimpleLedgerMapperTests.swift
@@ -3,68 +3,6 @@ import SwiftBeanCountModel
 import Wealthsimple
 import XCTest
 
-struct TestAsset: Asset {
-    var name: String = ""
-    var currency: String = ""
-    var symbol = ""
-    var type: AssetType = .currency
-}
-
-struct TestPositon: Position {
-    var accountId = ""
-    var asset: Asset = TestAsset()
-    var priceAmount = ""
-    var positionDate = Date()
-    var priceCurrency = ""
-    var quantity = ""
-
-    var assetSymbol: String {
-        get {
-            asset.symbol
-        }
-        set {
-            var newAsset = TestAsset()
-            newAsset.name = asset.name
-            newAsset.currency = asset.currency
-            newAsset.type = asset.type
-            newAsset.symbol = newValue
-            asset = newAsset
-        }
-    }
-
-    var assetType: AssetType {
-        get {
-            asset.type
-        }
-        set {
-            var newAsset = TestAsset()
-            newAsset.name = asset.name
-            newAsset.currency = asset.currency
-            newAsset.symbol = asset.symbol
-            newAsset.type = newValue
-            asset = newAsset
-        }
-    }
-}
-
-struct TestTransaction: Wealthsimple.Transaction {
-    var id = ""
-    var accountId = ""
-    var transactionType: TransactionType = .buy
-    var description = ""
-    var symbol = ""
-    var quantity = ""
-    var marketPriceAmount = ""
-    var marketPriceCurrency = ""
-    var marketValueAmount = ""
-    var marketValueCurrency = ""
-    var netCashAmount = ""
-    var netCashCurrency = ""
-    var fxRate = ""
-    var effectiveDate = Date()
-    var processDate = Date()
-}
-
 // swiftlint:disable:next type_body_length
 final class WealthsimpleLedgerMapperTests: XCTestCase {
